Cleaned simulation code

+++ src/main.rs
@@ -50,7 +50,7 @@ async fn main() -> Result<(), ()> {
             let time = Instant::now();
 
             if time.duration_since(last_world_step).as_secs_f32() >= WORLD_UPDATE_TIME {
-                let y = WORLD_SIZE as usize - 1;
+                let y = WORLD_SIZE - 1;
                 if world.get_cell(3, y) == Some(world::CellElement::Air) {
                     world.set_cell(3, y, world::CellElement::Sand(rand::thread_rng().gen_range(0.0..=2.0)));
                 }
